+++ benchmarks/prototype/moe_training/benchmark_scaled_grouped_mm_dq.py
@@ -23,10 +23,12 @@ from torchao.prototype.moe_training import _quantize_then_scaled_grouped_mm
 from torchao.prototype.moe_training.config import (
     FP8GroupedMMConfig,
     FP8GroupedMMRecipe,
+)
+from torchao.prototype.moe_training.utils import generate_jagged_offs
+from torchao.prototype.mx_formats.grouped_mm.config import (
     MXFP8GroupedMMConfig,
     MXFP8GroupedMMRecipe,
 )
-from torchao.prototype.moe_training.utils import generate_jagged_offs
 
 device = torch.device("cuda")
